reset_after_iteration defaults to False.
There are two reasons it is preferable for reset_after_iteration
to default to False.

The reason it was initially "False" is because it is very hard
to make counters that count up forever in a multi-process
system. But we'll just need to bite that bullet in CumulusCI
and Snowfakery because it's what users will want.