Remove uses of sets() which make iteration order unpredictable.
A user reported that in some cases her recipe loaded okay and in other cases it failed. The underlying cause of the failed loads is that there are two tables that had an implicit, not explicit, dependency. A picklist populated by CampaignMemberStatus had to be loaded before CampaignMembers could be.

We already have a way to force a load order, but my concern was that the intermittent behaviour made tracking down the source of the problem harder.

Also: the user's mental model was that the load order would follow the recipe.

AC:

This change ensures that the load order does actually follow the recipe except if a dependency can be inferred. Dependency order takes priority over recipe declaration order.