Allow synthetic references (#618)
A synthetic reference is when you know the 'id' of the thing
you want to reference and you want Snowfakery to treat it as
a reference. It is analogous to casting a pointer in C. Or
joining using an integer as a foreign key in a DB.

Currently this is not documented, because the internal id
mechanism isn't documented and may change. But it is useful
for some upcoming Salesforce-internal features and projects.
Once it's been tested there we can figure out how to make
it more widely available.

Probably in the long run, we will allow users to control their own
'id' mechanism on a per sObject basis. Then synthetic
references-by-id will make sense to them.